Removed item
I used Vue-CLI to install the application the first time.

I did encounter some issues, such as proxy not working the same way as the examples we were given. I found the solution for the proxy, but the scripts provided for importing the youtube-api did not work.

I tried remaking the app with Vite according to the examples provided in videos, but realized that Vite and Vue don't work together, and I've already spent a lot of time building the components. So I went back to my original frontend, and decided to look at the Group project how they solved the script issue. They did this by installing a "vue-youtube" dependency, which seems to be working for my project as well.

It was after this realization that I started to borrow more components from the group project as well.